requirement.txt
streamlit
osmnx
networkx
folium
matplotlib
requests
python-dotenv
